Move bodyParser into Router
